1-data-ingest-service/requirements.txt
ipykernel
functions-framework==3.*
requests
python-dotenv==1.0.0
pandas==2.0.3
pyarrow==12.0.1
google-cloud-bigquery==3.11.3
pandas-gbq==0.19.2
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.1
google-cloud-storage==2.10.0
